test: optimization of the test by mocking the user data
