+++ src/controllers/user/__tests__/userController.test.ts
@@ -2,59 +2,39 @@ import { Request, Response } from 'express';
 import { getUsers } from '../userController.ts';
 import * as userService from '../../../services/userService.ts';
 import handleResponse from '../../../utils/responseHandler.ts';
+import generateMockUsers from '../../../services/__mocks__/mockUsers.ts';
 import { User } from '../../../models/User/user.ts';
-import { encryptPassword } from '../../../utils/encryption.ts';
-import { v4 as uuidv4 } from 'uuid';
-import { UserRole } from '@prisma/client';
 
 jest.mock('../../../services/userService');
 jest.mock('../../../utils/responseHandler');
-jest.mock('../../../utils/logger');
 
 describe('User Controller - getUsers', () => {
   let req: Partial<Request>;
-  let res: Partial<Response>;
+  let mockRes: Partial<Response>;
+  let mockUsers: User[];
 
-  beforeEach(() => {
+  beforeEach(async () => {
+    mockUsers = await generateMockUsers();
+
     req = {};
-    res = {
+    mockRes = {
       status: jest.fn().mockReturnThis(),
       json: jest.fn(),
     };
     jest.clearAllMocks();
+
+    (handleResponse as jest.Mock).mockImplementationOnce((res, data) => {
+      res.status(data.status).json(data);
+    });
   });
 
   it('should return users with status 200', async () => {
-    const passwordHash = await encryptPassword('password123');
-    const oldPasswordHash = await encryptPassword('oldPassword123');
-
-    const mockUsers: User[] = [
-      {
-        id: uuidv4(),
-        userName: 'John Doe',
-        password: passwordHash,
-        oldPassword: oldPasswordHash,
-        role: UserRole.USER,
-        mail: 'john@example.com',
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      },
-      {
-        id: uuidv4(),
-        userName: 'Jane Smith',
-        password: passwordHash,
-        role: UserRole.EMPLOYEE,
-        mail: 'jane@example.com',
-        createdAt: new Date(),
-      },
-    ];
-
     (userService.getAllUsers as jest.Mock).mockResolvedValue(mockUsers);
 
-    await getUsers(req as Request, res as Response);
+    await getUsers(req as Request, mockRes as Response);
 
     expect(userService.getAllUsers).toHaveBeenCalledTimes(1);
-    expect(handleResponse).toHaveBeenCalledWith(res, {
+    expect(handleResponse).toHaveBeenCalledWith(mockRes, {
       data: mockUsers,
       message: 'Users retrieved successfully',
       status: 200,
